improved cmake build install targets + export cmake package config and version files merged libvorbis and libvorbisenc in libvorbis removed all vs projects and solutions in favor of cmake build examples with cmake (off by default) fixed seeking_example for msvc

+++ examples/seeking_example.c
@@ -29,7 +29,7 @@ void _verify(OggVorbis_File *ov,
              ogg_int64_t val,ogg_int64_t pcmval,double timeval,
              ogg_int64_t pcmlength,
              char *bigassbuffer){
-  off_t i;
+  size_t i;
   int j;
   long bread;
   char buffer[4096];
